requirements-tests.txt
-r requirements.txt
coverage
flake8
pytest
pytest-cov
pytest-django
